# Improve performance of liveness analysis in the register allocator (4.3X)