- Switch to using a `Vec` instead of `HashMap` since the key was an index anyways.
- When finding the successors, use a `label_to_index` hash map instead of looking for the index every time.
- Update how `modified` is computed and reduce the amount of things copied around.

I double checked that the bytecode size is unchanged for several small
and large designs so I know that the allocator is not less efficient in
that sense.

For https://github.com/sway-libs/concentrated-liquidity, the
improvements are as follows for the IR -> final bytecode step:
- Before this change: 42.72 seconds
- After this change: 9.88 seconds

**for a 4.3X improvement.**

> **Note: Combining this PR with https://github.com/FuelLabs/sway/pull/3898 and https://github.com/FuelLabs/sway/pull/3909, the total cost of converting the final IR to final bytecode goes down from ~263s to ~10s for a 26.3X improvement.**